Make the relation endpoint bidirectional.
Links currently have direction, they are inherently hierarchical. If the links for a record are requested via the /externalrecord/id/<id>/links/ endpoint we should also check to see if if a record has been linked as the "destination" link. For example, if we link (A) -> (B). We should see that A is linking B when B's links are requested.